+++ shared/core/types/actor.go
@@ -1,7 +1,5 @@
 package types
 
-// REFACTOR: Evaluate a way to define an interface for the actors which would enable to easily
-//           remove all the `switch actorType` statements in the codebase.
 func (a ActorType) GetName() string {
 	return ActorType_name[int32(a)]
 }