+++ shared/crypto/keypair.go
@@ -6,14 +6,14 @@ import (
 	"encoding/gob"
 )
 
-// Encoding is used to serialise the data to store the KeyPairs in the database
+// Gob Encoding is used to serialise the data to store the KeyPairs in the BadgerDB database
 func init() {
 	gob.Register(Ed25519PublicKey{})
 	gob.Register(ed25519.PublicKey{})
 	gob.Register(encKeyPair{})
 }
 
-// The KeyPair interface exposes functions relating to public and encrypted private key pairs
+// KeyPair interface exposes functions relating to public and encrypted private key pairs
 type KeyPair interface {
 	// Accessors
 	GetPublicKey() PublicKey
@@ -30,6 +30,9 @@ type KeyPair interface {
 	ExportString(passphrase string) (string, error)
 	ExportJSON(passphrase string) (string, error)
 
+	// Seed
+	GetSeed(passphrase string) ([]byte, error)
+
 	// Marshalling
 	Marshal() ([]byte, error)
 	Unmarshal([]byte) error
@@ -38,12 +41,13 @@ type KeyPair interface {
 var _ KeyPair = &encKeyPair{}
 
 // encKeyPair struct stores the public key and the passphrase encrypted private key
+// The encrypted private key is stored as a JSON string with the fields needed to decrypt the key
 type encKeyPair struct {
 	PublicKey     PublicKey
 	PrivKeyArmour string
 }
 
-// Generate a new KeyPair struct given the public key and armoured private key
+// newKeyPair Generates a new KeyPair interface given the public key and armoured private key
 func newKeyPair(pub PublicKey, priv string) KeyPair {
 	return &encKeyPair{
 		PublicKey:     pub,
@@ -51,27 +55,27 @@ func newKeyPair(pub PublicKey, priv string) KeyPair {
 	}
 }
 
-// Return an empty KeyPair interface
+// GetKeypair Returns an empty KeyPair interface
 func GetKeypair() KeyPair {
 	return &encKeyPair{}
 }
 
-// Return the public key
+// GetPublicKey Returns the public key
 func (kp encKeyPair) GetPublicKey() PublicKey {
 	return kp.PublicKey
 }
 
-// Return private key armoured string
+// GetPrivArmour Returns private key armoured string
 func (kp encKeyPair) GetPrivArmour() string {
 	return kp.PrivKeyArmour
 }
 
-// Return the byte slice address of the public key
+// GetAddressBytes Returns the byte slice address of the public key
 func (kp encKeyPair) GetAddressBytes() []byte {
 	return kp.PublicKey.Address().Bytes()
 }
 
-// Return the string address of the public key
+// GetAddressString Returns the string address of the public key
 func (kp encKeyPair) GetAddressString() string {
 	return kp.PublicKey.Address().String()
 }
@@ -81,7 +85,7 @@ func (kp encKeyPair) Unarmour(passphrase string) (PrivateKey, error) {
 	return unarmourDecryptPrivKey(kp.PrivKeyArmour, passphrase)
 }
 
-// Export Private Key String
+// ExportString Exports the private key as a raw string
 func (kp encKeyPair) ExportString(passphrase string) (string, error) {
 	privKey, err := unarmourDecryptPrivKey(kp.PrivKeyArmour, passphrase)
 	if err != nil {
@@ -90,7 +94,7 @@ func (kp encKeyPair) ExportString(passphrase string) (string, error) {
 	return privKey.String(), nil
 }
 
-// Export Private Key as armoured JSON string with fields to decrypt
+// ExportJSON Exports the private key as an armoured JSON string with fields to decrypt
 func (kp encKeyPair) ExportJSON(passphrase string) (string, error) {
 	_, err := unarmourDecryptPrivKey(kp.PrivKeyArmour, passphrase)
 	if err != nil {
@@ -99,7 +103,16 @@ func (kp encKeyPair) ExportJSON(passphrase string) (string, error) {
 	return kp.PrivKeyArmour, nil
 }
 
-// Marshal KeyPair into a []byte
+// GetSeed Returns the seed of the key
+func (kp encKeyPair) GetSeed(passphrase string) ([]byte, error) {
+	privKey, err := unarmourDecryptPrivKey(kp.PrivKeyArmour, passphrase)
+	if err != nil {
+		return []byte{}, err
+	}
+	return privKey.Seed(), nil
+}
+
+// Marshal Marshals the KeyPair into a []byte
 func (kp encKeyPair) Marshal() ([]byte, error) {
 	buf := new(bytes.Buffer)
 	enc := gob.NewEncoder(buf)
@@ -109,7 +122,7 @@ func (kp encKeyPair) Marshal() ([]byte, error) {
 	return buf.Bytes(), nil
 }
 
-// Unmarshal []byte into an encKeyPair struct
+// Unmarshal Unmarshals a []byte into an encKeyPair struct
 func (kp *encKeyPair) Unmarshal(bz []byte) error {
 	var keyPair encKeyPair
 	keypairBz := new(bytes.Buffer)
@@ -122,8 +135,8 @@ func (kp *encKeyPair) Unmarshal(bz []byte) error {
 	return nil
 }
 
-// Generate new private ED25519 key and encrypt and armour it as a string
-// Returns a KeyPair struct of the Public Key and Armoured String
+// CreateNewKey Generates a new private ED25519 key and encrypt and armour it as a string
+// Returns a KeyPair interface and error if any
 func CreateNewKey(passphrase, hint string) (KeyPair, error) {
 	privKey, err := GeneratePrivateKey()
 	if err != nil {
@@ -141,7 +154,8 @@ func CreateNewKey(passphrase, hint string) (KeyPair, error) {
 	return kp, nil
 }
 
-// Generate new KeyPair from the hex string provided, encrypt and armour it as a string
+// CreateNewKeyFromString Generates new KeyPair from the hex string provided, encrypt and armour it as a string
+// Returns a KeyPair interface and error if any
 func CreateNewKeyFromString(privStrHex, passphrase, hint string) (KeyPair, error) {
 	privKey, err := NewPrivateKey(privStrHex)
 	if err != nil {
@@ -159,7 +173,31 @@ func CreateNewKeyFromString(privStrHex, passphrase, hint string) (KeyPair, error
 	return kp, nil
 }
 
-// Create new KeyPair from the JSON encoded privStr
+// CreateNewKeyFromSeed Generates a new KeyPair from the seed provided
+// Returns a KeyPair interface and error if any
+func CreateNewKeyFromSeed(seed []byte, passphrase, hint string) (KeyPair, error) {
+	// Generate PrivateKey interface form secret key
+	reader := bytes.NewReader(seed)
+	privKey, err := GeneratePrivateKeyWithReader(reader)
+	if err != nil {
+		return nil, err
+	}
+
+	// Armour and encrypt private key into JSON string
+	armouredStr, err := encryptArmourPrivKey(privKey, passphrase, hint) // No passphrase or hint as they depend on the master key
+	if err != nil {
+		return nil, err
+	}
+
+	// Return KeyPair interface
+	return &encKeyPair{
+		PublicKey:     privKey.PublicKey(),
+		PrivKeyArmour: armouredStr,
+	}, nil
+}
+
+// ImportKeyFromJSON Creates new KeyPair from the JSON encoded privStr
+// Returns a KeyPair interface and error if any
 func ImportKeyFromJSON(jsonStr, passphrase string) (KeyPair, error) {
 	// Get Private Key from armouredStr
 	privKey, err := unarmourDecryptPrivKey(jsonStr, passphrase)