+++ p2p/CHANGELOG.md
@@ -7,10 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.30] - 2023-02-28
+## [0.0.0.31] - 2023-03-01
 
 - replace `consensus_port` with `port` in P2P config
 - update default P2P config `port` to from `8080` to `42069`
+
+## [0.0.0.30] - 2023-02-28
+
+- Renamed package names and parameters to reflect changes in the rest of the codebase
 
 ## [0.0.0.29] - 2023-02-24